Fix declaration INT32 & UINT32 for OS/2

+++ typesize.h
@@ -257,8 +257,8 @@ typedef unsigned char      UCHAR;               /*  1 byte */
 typedef signed   short int  INT16;              /*  2 byte */
 typedef unsigned short int UINT16;              /*  2 byte */
 /*#if !defined(__OS2_H__)*/
-typedef signed   long  int  INT32;              /*  4 byte */
-typedef unsigned long  int UINT32;              /*  4 byte */
+typedef signed   int        INT32;              /*  4 byte */
+typedef unsigned int       UINT32;              /*  4 byte */
 /*#endif*/
 /*  -------------------------------------------------------------------------- */
 #if !defined(__OS2_H__) && !defined(__NT__)  /*  os2.h and windows.h defines it already */
